+++ src/individual.cpp
@@ -3,7 +3,7 @@
 
 namespace genetic {
 
-/*static*/ std::function<int(individual const &)> individual::evaluation_function = &individual::eval_sum;
+/*static*/ std::function<float(individual const &)> individual::evaluation_function = &individual::eval_sum;
 /*static*/ std::function<void(individual *, individual *)> individual::mating_function = &individual::two_point_crossover;
 /*static*/ std::function<void(individual &)> individual::mutation_function =
       std::bind(&individual::shuffle_indexes, std::placeholders::_1, static_cast<float>(0.05));
@@ -88,7 +88,7 @@ void individual::shuffle_indexes(float mutation_rate) {
    }
 }
 
-int individual::evaluate() {
+float individual::evaluate() {
    if (!valid_fitness) {
       fitness = evaluation_function(*this);
       valid_fitness = true;
@@ -96,11 +96,11 @@ int individual::evaluate() {
    return fitness;
 }
 
-/*static*/ int individual::eval_sum(individual const & ind) {
-   return static_cast<int>(std::accumulate(ind.attributes.begin(), ind.attributes.end(), 0.0,
-                                           [] (double current_sum, attribute const & attr) -> double {
-                                              return current_sum + static_cast<double>(attr);
-                                           }));
+/*static*/ float individual::eval_sum(individual const & ind) {
+   return std::accumulate(ind.attributes.begin(), ind.attributes.end(), 0.0f,
+                                           [] (float current_sum, attribute const & attr) -> float {
+                                              return current_sum + static_cast<float>(attr);
+                                           });
 }
 
 bool individual::operator<(individual const & other) const {
